Move question to own app
