+++ backend/experiment/rules/huang_2022.py
@@ -7,11 +7,11 @@ from django.conf import settings
 from experiment.actions import HTML, Final, Explainer, Step, Consent, Redirect, Playlist, Trial
 from experiment.actions.form import BooleanQuestion, Form
 from experiment.actions.playback import Autoplay
-from experiment.questions.demographics import EXTRA_DEMOGRAPHICS
-from experiment.questions.goldsmiths import MSI_ALL, MSI_OTHER
-from experiment.questions.other import OTHER
-from experiment.questions.utils import question_by_key
-from experiment.questions import QUESTION_GROUPS
+from question.demographics import EXTRA_DEMOGRAPHICS
+from question.goldsmiths import MSI_ALL, MSI_OTHER
+from question.other import OTHER
+from question.utils import question_by_key
+from question.questions import QUESTION_GROUPS
 from experiment.actions.styles import STYLE_BOOLEAN_NEGATIVE_FIRST
 from result.utils import prepare_result
 from .hooked import Hooked
