+++ backend/question/fixtures/question.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "dgf_generation",
     "fields": {
       "question": "When were you born?",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "dgf_gender_identity",
     "fields": {
       "question": "With which gender do you currently most identify?",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "P01_1",
     "fields": {
       "question": "Can you clap in time with a musical beat?",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "P01_2",
     "fields": {
       "question": "I can tap my foot in time with the beat of the music I hear.",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "P01_3",
     "fields": {
       "question": "When listening to music, can you move in time with the beat?",
@@ -40,7 +40,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "msi_01_music_activities",
     "fields": {
       "question": "I spend a lot of my free time doing music-related activities.",
@@ -48,7 +48,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "msi_03_writing",
     "fields": {
       "question": "I enjoy writing about music, for example on blogs and forums.",
@@ -56,7 +56,7 @@
     }
   },
   {
-    "model": "experiment.question",
+    "model": "question.question",
     "pk": "msi_08_intrigued_styles",
     "fields": {
       "question": "I’m intrigued by musical styles I’m not familiar with and want to find out more.",
