+++ backend/question/fixtures/questionseries.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "experiment.questionseries",
+    "model": "question.questionseries",
     "pk": 1,
     "fields": {
       "experiment": 14,
@@ -9,7 +9,7 @@
     }
   },
   {
-    "model": "experiment.questionseries",
+    "model": "question.questionseries",
     "pk": 2,
     "fields": {
       "experiment": 20,
